# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VK Bot API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type VkCallbackEventObject = { [key: string]: unknown };

export interface VkCallbackEvent {
  type: string;
  object?: VkCallbackEventObject;
  group_id?: number;
  /** @nullable */
  secret?: string | null;
}

export interface Command {
  id: number;
  trigger: string;
  response: string;
  /** @nullable */
  description?: string | null;
  isActive: boolean;
  usageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCommandBody {
  trigger: string;
  response: string;
  /** @nullable */
  description?: string | null;
  isActive?: boolean;
}

export interface UpdateCommandBody {
  trigger?: string;
  response?: string;
  /** @nullable */
  description?: string | null;
  isActive?: boolean;
}

export interface BotSettings {
  communityName: string;
  welcomeMessage: string;
  unknownCommandMessage: string;
  commandPrefix: string;
  isActive: boolean;
}

export interface CommandUsage {
  trigger: string;
  usageCount: number;
}

export interface BotStats {
  totalMessages: number;
  commandsUsed: number;
  activeCommands: number;
  totalCommands: number;
  topCommands: CommandUsage[];
}

export interface MessageLog {
  id: number;
  userId: string;
  message: string;
  /** @nullable */
  commandTrigger?: string | null;
  /** @nullable */
  responseSent?: string | null;
  createdAt: string;
}

export type ListMessageLogsParams = {
  limit?: number;
};