# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VK Bot API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Receives events from VK
 * @summary VK Callback API webhook
 */
export const VkCallbackBody = zod.object({
  type: zod.string(),
  object: zod.object({}).passthrough().optional(),
  group_id: zod.number().optional(),
  secret: zod.string().nullish(),
});

/**
 * @summary List all commands
 */
export const ListCommandsResponseItem = zod.object({
  id: zod.number(),
  trigger: zod.string(),
  response: zod.string(),
  description: zod.string().nullish(),
  isActive: zod.boolean(),
  usageCount: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListCommandsResponse = zod.array(ListCommandsResponseItem);

/**
 * @summary Create a new command
 */
export const CreateCommandBody = zod.object({
  trigger: zod.string(),
  response: zod.string(),
  description: zod.string().nullish(),
  isActive: zod.boolean().optional(),
});

/**
 * @summary Get a command by ID
 */
export const GetCommandParams = zod.object({
  id: zod.coerce.number(),
});

export const GetCommandResponse = zod.object({
  id: zod.number(),
  trigger: zod.string(),
  response: zod.string(),
  description: zod.string().nullish(),
  isActive: zod.boolean(),
  usageCount: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a command
 */
export const UpdateCommandParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateCommandBody = zod.object({
  trigger: zod.string().optional(),
  response: zod.string().optional(),
  description: zod.string().nullish(),
  isActive: zod.boolean().optional(),
});

export const UpdateCommandResponse = zod.object({
  id: zod.number(),
  trigger: zod.string(),
  response: zod.string(),
  description: zod.string().nullish(),
  isActive: zod.boolean(),
  usageCount: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a command
 */
export const DeleteCommandParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get bot settings
 */
export const GetBotSettingsResponse = zod.object({
  communityName: zod.string(),
  welcomeMessage: zod.string(),
  unknownCommandMessage: zod.string(),
  commandPrefix: zod.string(),
  isActive: zod.boolean(),
});

/**
 * @summary Update bot settings
 */
export const UpdateBotSettingsBody = zod.object({
  communityName: zod.string(),
  welcomeMessage: zod.string(),
  unknownCommandMessage: zod.string(),
  commandPrefix: zod.string(),
  isActive: zod.boolean(),
});

export const UpdateBotSettingsResponse = zod.object({
  communityName: zod.string(),
  welcomeMessage: zod.string(),
  unknownCommandMessage: zod.string(),
  commandPrefix: zod.string(),
  isActive: zod.boolean(),
});

/**
 * @summary Get bot statistics
 */
export const GetBotStatsResponse = zod.object({
  totalMessages: zod.number(),
  commandsUsed: zod.number(),
  activeCommands: zod.number(),
  totalCommands: zod.number(),
  topCommands: zod.array(
    zod.object({
      trigger: zod.string(),
      usageCount: zod.number(),
    }),
  ),
});

/**
 * @summary Get recent message logs
 */
export const ListMessageLogsQueryParams = zod.object({
  limit: zod.coerce.number().optional(),
});

export const ListMessageLogsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  message: zod.string(),
  commandTrigger: zod.string().nullish(),
  responseSent: zod.string().nullish(),
  createdAt: zod.string(),
});
export const ListMessageLogsResponse = zod.array(ListMessageLogsResponseItem);