# lib/api-zod/src/generated/types/botSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VK Bot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BotSettings {
  communityName: string;
  welcomeMessage: string;
  unknownCommandMessage: string;
  commandPrefix: string;
  isActive: boolean;
}