# lib/api-zod/src/generated/types/botStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VK Bot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CommandUsage } from "./commandUsage";

export interface BotStats {
  totalMessages: number;
  commandsUsed: number;
  activeCommands: number;
  totalCommands: number;
  topCommands: CommandUsage[];
}