# lib/api-zod/src/generated/types/commandUsage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VK Bot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CommandUsage {
  trigger: string;
  usageCount: number;
}