# lib/api-zod/src/generated/types/messageLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VK Bot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MessageLog {
  id: number;
  userId: string;
  message: string;
  /** @nullable */
  commandTrigger?: string | null;
  /** @nullable */
  responseSent?: string | null;
  createdAt: string;
}