# lib/api-zod/src/generated/types/vkCallbackEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VK Bot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VkCallbackEventObject } from "./vkCallbackEventObject";

export interface VkCallbackEvent {
  type: string;
  object?: VkCallbackEventObject;
  group_id?: number;
  /** @nullable */
  secret?: string | null;
}